streamline tests
